Bugfix: race between StreamBegin and StreamSync.
kvfeed will spawn a vbucket-routine when it sees StreamBegin for that vbucket.
Immediately after that it was posting the StreamBegin Event to newly spawned
routine.

Race happens when kvfeed is switched out after it spawned the routine but
before posting the event. The new vbucket-routine gets a chance to run and
starts emitting the heartbeat for every 5ms. Leading to StreamSync before
StreamBegin.

- feed does not send vbmap now. StreamBegin is the sole event that opens the
  path for a vbucket.

Change-Id: Ic93c72f087d336af9d587f85b0e2e949b2f739b6
Reviewed-on: http://review.couchbase.org/39049
Reviewed-by: Deepkaran Salooja <deepkaran.couchbase@gmail.com>
Reviewed-by: Pratap Chakravarthy <prataprc@couchbase.com>
Tested-by: Pratap Chakravarthy <prataprc@couchbase.com>