Call newly added forestdb destroy api on Drop Index
When an index is dropped, the physical files need to be
cleaned up as well. ForestDB now has a destroy api to do that.
That api needs to be called for all slices of an index.

There is a corresponding goforestdb commit, which needs to be
merged along with this.

Change-Id: I489ce815315eb6e5839917fbc3dbf15c1d1e488d
Reviewed-on: http://review.couchbase.org/43414
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Deepkaran Salooja <deepkaran.couchbase@gmail.com>
Tested-by: Deepkaran Salooja <deepkaran.couchbase@gmail.com>