storage_manager: Notify waiters only if they belong to update bucket
The storage manager was updating timestamps for all indexes belonging
to the current indexer when any of the buckets have a new timestamp for
snapshot creation. This results in updating index-timestamp
with with wrong ts for which no snapshot exist.

Fixed storage manager to only update indexes belonging to the currently
processed bucket.

Change-Id: I0300c985cb3bc20c90081923a6ea133490e03bec
Reviewed-on: http://review.couchbase.org/43545
Reviewed-by: Pratap Chakravarthy <prataprc@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Tested-by: Sarath Lakshman <sarath@couchbase.com>