granular protobuf dependency and cleanup.
protobuf/ package is split into three sub-packages
    projector/ - for projector adminport messages
    data/ - for mutation path from projector to indexer
    query/ - for querying index.
This will reduce this risk of circular dependency.

I have cleaned up most of the un-used protobuf messages which were originally
added to capture design discussion on 2i. Now that we have manager and indexer
evolving with a code-base, these messages and associated descriptions are not
necessary.

Change-Id: If7dd4763fcef50ac355e4ba53685cf1f4aadf960
Reviewed-on: http://review.couchbase.org/43676
Reviewed-by: Pratap Chakravarthy <prataprc@couchbase.com>
Tested-by: Pratap Chakravarthy <prataprc@couchbase.com>