Reduce log layers
Currently, there are 3 different logging layers. This is obviously
not going to work in production. Reducing it to 2 for now. Soon,
we will need to decide between common/debug.go and go's log package
and use only one. Not making that refactor yet, as it needs to become
an independent package due to dependency issues.

Change-Id: I24bc152fc41ebee9b59ec1f9f51c3ae5b9c1afc6