KVSender Improvements for Repair Stream
- KVSender uses the projector idempotent behaviour for failed
requests.
- KVSender now does async processing of requests and result is
supplied on a response channel. This allows for simultaneous requests
for {stream, bucket, vbucket}.
- Bookkeeping for ActiveTs returned in MutationTopicResponse to
determine the completion of request.
- RestartVbuckets can determine the list of affected nodes for
a set of vbs and target the request to only those nodes.
- KVSender is now stateless to allow async processing. Bookkeeping
for Instance per {stream, bucket} is removed. The caller should do
this bookkeeping.
- DeleteBucket is now a explicit call for caller. As there is no
bookkeeping now, KVSender cannot determine when the last instance
for a bucket is deleted. Caller needs to call DeleteBucket.
- All requests now have a stop channel to abort the request
- RepairEndpoint is archived as it cannot be used without a
catchup mechanism
- Added lock for cluster_info usage

Change-Id: I2fced5f3e4d0c954fbb9834c7800ec9fbd901199