Make logging uniform. Rename UPR to DCP.
It is important that logging has a consistent format and can be
centrally controlled. This is not possible if we use multiple log
packages. So move everything over to a single logging location.
Due to dependencies, it is necessary to split logging out of common
package.

For merging convenience, this also includes MB-13059 to rename
UPR to DCP as decided during 3.0 days.

Change-Id: I948ccaa9757015696e2ecaa83ab7709573cb1a91