Import goforestdb as subtree repo
We need to add debugging and locking to goforestdb to
diagnose problems in how indexing uses goforestdb. But
these changes are not appropriate for main repo.

So copying it over temporarily. Will also revert the
ugly logging added recently for this purpose in the main
repo as it can be done in the copy here now.

Change-Id: I9d8ec5f5372ba39644f4814201689d00fb7dfefe