MB-100: Add ability to log pprof data periodically
We're seeing hangs when building indexes. But it's not always
possible to run pprof when the hang occurs as the environment
may be inaccessible or have gone away. So log the goroutine
profiles every once in a while to log file at DEBUG level.

Also add authentication for cbindex config option

Change-Id: I7350ed3a32296df7edac5bb958eb9c26a77c9ef7