MB-13823: CREATE INDEX hangs.
IMHO the bug can happen when back-channel for feed is populated for
shutdownVbuckes but unfortunately the feed's gen-server does not
get a chance to run during the timeout period leading to mangled
behaviour. When run with 2 nodes in a VM, the system had to deal
with host scheduler, guest scheduler and finally golang scheduler
and with TestMultipleBucket there can be upto 1024 * 4 go-routines.

Following fixes should make projector more resilient
to this behaviour.

- un-avoidable and valid race exists between
  feed.shutdownVbucketRequest() and async StreamEnd that will
  clean up stream-data-structures along its path downstream.
  * log warning and pretend everything is normal.
  * only shutdown active-vbuckets, though this does not avoid
    the above mentioned race.
- race condition between DcpFeed.Close() and DcpFeed.doReceive().
- log the error if feed.waitStreamRequest() or feed.waitStreamEnd()
  returns an error.

Change-Id: I6802ce2fc8970e0caf1fdb846920ec4fc4e98fc0