MB-13885 queryport: Fix client to use DefnID of equivalent index
Based on the load distribution, gsi client may choose to use a
different index with same index defintion on a different node to
perform scan operation. But, even though indexer address was chosen
correctly, the chosen index's defnID was not passed in the scan
request. Fixed client to use new index defnID instead of old defnID.

Change-Id: Ib6ca2b1fb1873a760bc5d51d404a3324e55c2975