MB-14090: handle ns-server rewrites server node name on VM (single node cluster)
When using deployment plan, the indexer IP must be provided.  When running
on a single node cluster on VM, ns-server rewrites the server node name
to real IP address, but GsiClient does not get notified on the name change.
With this fix, whenever GsiClient/MetadataProvider discovers a potential
stale node name, it will ask ns-server to refresh indexer list from
ns-server once, with the expectation that ns-server will be able
to provide the "real" server node name during refresh.

Change-Id: I4e1a48875f98175caa495a1efaa3290ce73b34cb