MB-14115: fix to add a new config param.
When back-channel was introduced for feed book-keeping in projector,
the channel size was same as feed-channel size, which was 100.
Later to avoid back-channel path from blocking, we increased the channel
size to 10000, but at this point we should have created a new
config param for back-channel.

Due to recent mprof optimization, the channel size for feed's control
path was reduced back to 100.

Now added a new configuration param - backChanSize whose size is
10000 by default.

Change-Id: I52c6c8738f1ef5e4ce5e1fd6002908b80bbebc2b