MB-13590 Indexer to have its own timer/sync msg mechanism
With limiting of mutation queue length, it is not always going to
be possible for indexer to keep listening to sync messages from
projector. Both functionalities being served by sync(updation of HWT
and providing a timer facility to generate stability timestamps periodically)
are now managed by Indexer on its own.

Change-Id: I9547670540abee6f47c12969e0b4aa61db6942dd