MB-14957: implement nuke-switch for indefinitely blocked feeds.
Indexer might some time hang due to indefinitely blocked call to projector.
To avoid this, we propose two solutions.

a) there is a routine that periodically checks for the status of all
   active feeds. we can add a timeout to status-check calls (say 5-10 minutes)
   and if feed does not respond a healthy status within the timeout period,
   we can do a force shutdown of that feed and clean up the local data-structure.
b) this a variant of (a) where if feed does not respond a healthy status
   within the timeout period, we can crash the projector and have it restarted
   by baby-sitter process.

this patch implements option(a).

Change-Id: I8b42702fb0686023ba7fbf0ac2f2013cd7832e18