MB-15241 indexer: Avoid stats update for deleted buckets
All the indexes for a bucket can get deleted and still flusher
might be started drain the mutations from the queue. Since, bucket
already dissapeared from the indexer, stats object bucket list
won't have a corresponding entry. We should not assume that
bucket stats object is always available if flusher is invoked for
a corresponding bucket.

Change-Id: I0aba9797571c5c3060e4f524d4d3485e218ac81f