MB-15465 Fix indexer loop blocking in compaction daemon
Indexer hangs when a new config is posted to compaction daemon
when it is executing compaction. Indexer config update is synchronous
in nature and requires all components to respond immediately. Since
compaction daemon waits for currently running compaction routine to
finish, it blocks the entire indexer loop. Fixed compaction daemon
to reset the config asynchronously.

Change-Id: I5046005b51c8a1df822e18c5364c1cedf268ccd9