MB-15567 Cleanup RestartTs from Timekeeper in Recovery
Now the streams are only closed and reopened in case there is
a rollback. All other recovery procedures work without any stream
close. So timekeeper doesn't need to retain the restartTs and
it is guaranteed to get the restartTs in stream open message.

The current bug was happening because the restartTs used to be retained
in rollback based recovery. The recovery itself would get aborted due to
all indexes being dropped. When next stream open came, it was using the
old restartTs leading to incorrect flush behaviour.

Change-Id: Ic6d10d9fca7f8e0b74972fb8781625400093dec7