MB-15626 Generate TS/Snapshot on Vbuuid Change
New TS needs to be generated for vbuuid change as stale=false scans
can only succeed if all vbuuids are latest. Also if there are no docs
in a vbucket and its stream begin arrives later than all mutations,
there will be no TS with that vbuuid.

This also fixes the case of ep-engine restart, as there will be a
new stream begin. Even if there are no mutations, stream begin will
cause new TS/snapshot to be generated.

Change-Id: I3b9c6062cbd03e3447fd4d60888d3fb40ed644c3