MB-15611 client: Add on-the-fly request batching for get-seqnos calls
We have only single dcp connection per bucket per node for sending
get-seqno requests. Since many client workers need to use the same
connection, we could batch get-seqno requests and issue only single
get-seqno request to the EP-Engine and broadcast the same response
to all the requests in a batch. Batching is done on-the-fly to avoid
any additional latency due to batching.

Change-Id: Ia8bc59eacca921daaffad8bb13626f780ec95cb0