MB-15645: fold error paths.
N1QL expects error via a channel, which was getting posted via callback
right when the error is detected (that is, while reading the message
from socket)

This happens asynchronously with the same error returned back to
client SDK, which implements the retry logic.

this patch handles,

1. fold both the error path so that error can surface up to N1QL after
retry loop fails.
2. avoid retry when error happens after a partial scan.

Change-Id: Ib518f0f805e5993a16fe1d31d589d4c6f4090d37