MB-15679 Skip mutations in stream_reader without StreamBegin
This will take care of the residual mutations that can come
when projector returns success for DelBuckets but mutations
in the endpoint queue are going to be sent to dataport server.

This happens typically in recovery for multiple bucket when one
bucket is stopped(using DelBuckets) and then restarted(using MTR).

Change-Id: I128ba7a66d4937f35def93c95ff26b00236dbfcd