MB-16154: Create Breakpad Minidump file upon fatal forestdb error
Make use of the new SetFatalErrorCallback() forestdb API to trigger
the creation of a Breakpad minidump when forestdb encounters a fatal
error.

Implemented by using breakpad_wrapper shared library which exposes a
simple C API to Breakpad (C++). This allows Golang (specifically indexer)
to setup Breakpad and manually trigger the creation of Minidumps. This
nterface is then wired up with the forestdb SetFatalErrorCallback() API
such that we register Breakpads' WriteMinidump() function as the fatal
error callback.

Minidump files will be written to the path specified by the
'minidumpDir' argument, or the storageDir if minidumpDir is not
specified.

Note: We *cannot* use Breakpad's handling of fatal errors as Golang
      gets very upset if other code attempts to change the signal
      handlers for SEGV, SIGBUS or SIGFPE - which is exactly what
      Breakpad will do if configured to handle fatal signals.

Change-Id: I5408b597e4d1073ab78153ccb4f3d2091bb8160e