MB-16275: synchronously close connection.
since gen-server returns immediately, there is a case where
the downstream routines are closed and the dcp's output channel
is not drained.

when the connection routine, as part of cleanup, publishes the
stream-end on the dcp's output channel it might find it full
and block - hence the hang.

solution is to synchronously close the upstream routines
before returning.

Change-Id: Ib05dc8c394dd1e7757acbf30b24761c16da86432