MB-16275: synchronously close connection.
since gen-server returns immediately, there is a case where
the downstream routines are closed and the dcp's output channel
is not drained.

when the connection routine, as part of cleanup, publishes the
stream-end on the dcp's output channel it might find it full
and block - hence the hang.

solution is to synchronously close the upstream routines
before returning.

Change-Id: Iff620724bff7755124df9ecc9d8608e03ef3e1a3