MB-17226: Backport MB-17009 - Do not callback from fdb to go
This fixes the cherry pick of MB-17009 (ac51957) to work properly
with Go 1.4 on Windows, where compiling .c modules doesn't work.

MB-17009: Do not callback from forestdb into go

As forestdb has threads originating from go and also threads it
starts up by itself, the execution environment is a mix of cgo
thread and pthreads. It is not possible to predict which thread
will execute callbacks. So it is best to keep all callbacks in C.

Change-Id: I137d04d8f5f8a31cbf302e9df0aef0402b156f53