MB-17849 Directly enqueue mutations to queue
buffered workers have not provided any performance gain as
the main stream reader loop is single threaded. If worker
is holding onto a mutation waiting for enqueue, it can lead
to a hang. It is better to let go of workers. Next step would
be to make the main stream reader loop concurrent.

Change-Id: I33ac3206ff61ca037fe513941c10aa1451e1828e