MB-19005: delete feed on automatic shutdown.
Right now the feed is not deleted when there is an automatic shutdown
due to downstream socket failures.
This was left as such to make Feed{} object agnostic of Projector{}.
Looks like that is not a right thing to do anyway.
Later, if need be, we can abstract the *Projector parameter into
an interface.

Additionally, runtime.GC() is called when ever a valid feed is
deleted from projector. This might help is reclaiming the garbage
earlier before reachable heap get adjusted to higher value.

Change-Id: I331c6df083dd116a119faa40bac53f9f4b7ce394