MB-19064 : Fix rollback to 0 when some VBs are not rolled back
When Rollback is received by indexer, some of the vbuckets
are not present in rollbackTS which led to those VB seq num
becoming 0 leading to rollback to 0. Fix is to do union of activeTs
with rollbackTs in kv_sender. Also fixed the condition which
led to only latest persisted snapshot being compared in GetOlderThanTS
method

Change-Id: Ia80cba1c08341d106b38806e1e99db5863430b11