MB-19426 Release queue map lock after sync worker is done
This is a RLock and should be ok to hold on to till sync worker
is done. The only case this will slow down is the rollback case or
when index for a bucket is first created/dropped, which is not that
often.

The problem by not holding onto the queue lock can be that there is a
small window in rollback when sync worker can see the old filter.

Change-Id: I3de840325d33c61601a6877b528a7fa5c124641b