MB-19499 memdb: Close() should wait until all iterators have exited
The memdb frees all the nodes before all the iterators have finished.
During bucket flush or rollback, we close the memdb instance immediately
and creates a new instance. But, there could be live iterators for the
old instance and it may accees free nodes. This leads to crash.
Implemented waiting in the memdb.Close() until all snapshots have been
closed.

Change-Id: If747d575b8fb10de525b20ce47c0eeb07acd11d1