MB-19524 copy index key in stream reader
Right now we are keeping the reference to protobuf key. But
this leads to the protobuf garbage not being GCed till the mutations
go to storage(mutations sitting in mutation queue, flush queue etc).

As we have multiple stream workers now, we can afford to copy the key
as this will be done in parallel and let the protobuf garbage be freed
early.

Change-Id: I99aefc4c5b39e0dbfc5489f6cd4b6a63c9ea43a7