MB-19531 moi/array: Fix crash due to SMR algorithm violation
Currently, the back index update is performed after all node
deletes are done. The back index does not maintains a pointer
from crc32(docid) to node. When an update is performed, first
it needs to locate the current entry by validating the docid
from the corresponding node. Since the node is already deleted,
SMR may free it up. Hence, access to the freed node during update
will access an invalid address and crash.

Change-Id: I99aa5ad00749051d14100951c87d8dfd322bb40e