MB-20116 Fix race condition during pause in stream reader
stream reader needs to acquire the indexer state before it
starts to process a mutation. Otherwise it is possible that
a mutation doesn't get enqueued to the mutation queue(due to indexer
moving to paused state) but its sync message gets processed by
timekeeper and a TS gets generated. Such a TS will never complete
and keep waiting.

Change-Id: I9bc6bcf6f7220af2d28250e241451297cedba723