MB-20872 Handle alignment issue due to partial snapshot
Timekeeper disables alignment for timestamp if partial
snapshots are seen. However, there can be a sequence of events
which can lead to alignment issues(see comment on MB-20872 for details).
Timekeeper already has logic to make sure TS seqnums are monotonically
increasing. But that logic gets skipped if alignment is disabled for ts.
This change makes sure that logic to make ts monotonic is always called.

Change-Id: I7f3b6e0e4e94c60b833e87e136f6ae89eefdc71d