MB-20847 Fix compaction for non mutating indexes
Some indexes may not mutate with KV mutations if
the indexed field doesn't change or if its a partial
index(where clause). This will cause the main and
back kvstore seqnums to remain the same.

The check needs to be on default kvstore's seqnum as
that is guaranteed to increment with every commit.

Change-Id: I533da5d29ecc42fd3a67f4560ad27077e7ac2ab4