CBIDXT-403: Select items to return in scans
Support Index projection as part of MultiScan
Index Projection has a list of keys mapping
composite element positions to indicate subset
keys to project. It also has a flag "PrimaryKey"
to indiciate whether primary key has to be projected
or not.

Distinct is NOT supported when projection is specified
for secondary keys.

Change-Id: I2c77ae4c27eac8486b74ecfffce08a6f6401f90d