MB-22409 Increase buffer length of all encode buffers
Due to incorrect check for key lengths in code, extra buffer space
is required to make sure there is no panic while writing to encode buffers.

The proper fix to correct the checks in all places is much involved and risky
at this point. This is a safe fix to make sure encode buffers always have
enough space.

Change-Id: Ifb15faa75a12b70add481c475aca3daae2919592