MB-23463: Fix bug in prefix sort of scan index keys
If prefix comparison of two IndexKeys results in 0
and if the keys are of same length, it implies they
are exact same values. In this case if first IndexKey (x)
is low and second (y) is high, then x is < y as high key
is more inclusive making it higher on sort order.

Change-Id: Id0bf78e39967032167f48b9b6fa69ea67ce082de