MB-23828 plasma: Fix index drop stall during concurrent index scan
A goroutine which holds the last reference processes the slice destroy
operation. But, if the last routine is a scan routine, it will hold an
index reader context and try to close the slice. The slice close operation
tries to reclaim all reader contexts and hence it ends up in a deadlock.

Since the refcounting mechanism avoids possibility for leaking contexts,
it we do not need to wait for reclaimation of reader contexts

Change-Id: I003321dde745d230c6acbf8d462ef94e287cdb0d