MB-26699 [Backport 5.1.0] Validate checksum only for incoming server packets
For the outgoing packets(i.e. scan response), disable adding checksum.
In case of gsi client, it caches both the connection and transport packet.
If new server sends checksum to old client, the client is going to retain
the checksum in the packet. Whenever the packet gets reused from the pool,
the stale checksum will be sent to the server causing mismatch.

This mixed mode problem can only be solved once the complete fix is done(
which is to only send the new packet format once the cluster has fully upgraded).

Change-Id: Id5270a947831b9f7b9347d42e7471adbc72d431d