MB-27345 - Proper error handling should be introduced in the call to Send..
In package transport...
The purpose of the patch is that it should return a proper
error message (like io.EOF in case connection of reset), instead of
returning ErrorPacketWrite. Note that, if EOF is returned GsiClient will
follow Evict-retry path, otherwise it will retry one more time and fail
the request.

Change-Id: I0e4f5f74a0aaad3d2e3553648e757ab98b1a48e9