MB-28305: Fix primary scan check in client for API2
This code change contains two fixes for API2 primary scans
serialization in scan_client:
1. There is no nil check needed before calling curePrimary as
   a nil value passed from n1ql for API2 Range low/high indicates
   it is a json null
2. A request contains multiple scans. Currently if curePrimary
   fails any of the scan, we are returning from MultiScanPrimary or
   MultiScanCountPrimary. Instead return only if all scans are
   invalidated by curePrimary.

Also this fix is only for API2. For API3, since it can include
count as aggregate, the client needs to send the request to server
even if it is invalid range. The fix for API3 is adifferent change set.

Change-Id: I5c75626fa30fec9fe64d71eac188d5541cb0578a