MB-25315 identify mutations from first DCP snapshot
A new per vbucket flag is maintained in each stream worker to mark
if first snapshot is being processed.  Any stream restart/repair/failure
will turn the flag off, even if the mutations are coming from the
first snapshot. The only exception is Rollback To Zero(as it will reset filter).

Change-Id: I855a774e1eff4cad85f4ed727da37b5e95f0e898