MB-28838: Fix comparison method used in primary key range processing
Currently bytes.Compare is used in a pre-processing step to compare
primary index low and high. When either low and high is empty string
and other is Min/Max, in this case bytes.Compare return them as equal.

The right method for comparison is CompareIndexKey method on PrimaryKey
type which takes care of Min/Max.

Change-Id: I30a3ce35cf90a5b1b0c9ea974fa944412caa5b1f