MB-26307: moi - Throttle number of StoreToDisk by numCPU
When number of indexes > number of cores, the large number of
persistence go routines ends up causing CPU spikes which results in
increased query latency.
To solve this have a buffered channel = number of CPU cores and only
have as many persistence go routines as number of cores.
So the max number of go routines spawned = numCPU * concurrency
Add ability to dynamically tune the number of persisters

Change-Id: If5c66d8f7ee86498fb1667f800d38f6679ada789