MB-28139: Don't fail indexer bootstrap if index storage is corrupted.
I.
Detect the storage corruption and handle it gracefully. If storage
is corrupt for an index (or a partition of an index):
(1) Don't make it available
(2) Cleanup its data
(3) Notify the administrator about the corruption and cleanup.

II.
This allows availability of non-corrupted indices.

III.
Allow DeleteOrPruneIndexInstance during bootstrap

IV.
For MOI, detect disk corruption in NewStorageManager.
(1) If latest snapshot is corrupted, check for previous snapshot.
Perform this until all snapshots are attempted.
(2) If a corrupt snapshot is found, delete its directory.
(3) Open with any non-corrupted snapshot that is found
- in chronologically descending order.
(4) If all snapshots are corrupted: Create error file on
persistent storage and let indexer exit. Once the indexer
restarts, check for presence of error file during NewSlice. If there
was a corruption error, cleanup the index partition.

Change-Id: Iea73963b2cd06c8941fc1ca21a4d58c97ea1cc3c