MB-30072: Respond to DCP_NOOP immediately, in projector dcp consumer
DCP_NOOP message is used for keeping alive the connection between
dcp producer and dcp consumer. If DCP_NOOP is added to receive channel,
it will have to wait for all other events in the receive channel to
be processed, before consumer can repond to it. If this takes more
time than dcp noop interval, producer will wrongly close the connection.
To avoid this, respond to DCP_NOOP immediately, without adding
it to receive channel.

Change-Id: I0c4d540e48ad4ecfbd4d380dd209499f09ff626b