MB-30382 plasma: Stop mutation queue workers before resetting buffers
When the indexer processes drop index, it flushes the mutations to the
slice and issues drop index. When indexer operates in NO_SNAP mode (due
to large number of mutations in the snapshot or unable keep up with the
drain rate), a new snapshot is not created before the slice.Close() is
called. Every snapshot creation waits for current slice queue to drain.
But in NO_SNAP case, the residue mutations will be processed by the
slice mutation workers while slice.Close() is processed. Currently, the
buffers are reset before stopping the workers leading to plasma inmemory
pages getting corrupted.

Change-Id: I8a378f73b9fcf32f757b20cdec557b58b86b1870