MB-30788: Introduce recovery in CollectSeqnos
The observed panic is seen only if kvfeeds passed to CollectSeqnos
is an empty map. We don't expect this to happen as this possibility
is safeguarded by numVbs check in addDBSbucket. Once the object of
vbSeqnosReader is created, there is no code that changes the value of
vbSeqnosReader.kvfeeds.

This code change avoids panic in CollecSeqnos. This is just a safeguard
which ensures progress as the retry happens with updated clusterinfo.

Change-Id: Ibdb0cc81530505e73ca9b1a6bfeec456dd5a3f72