MB-32527: Fix for building with Go 1.10+
When building with Go 1.10.4 or 1.11, cbindex fails to compile complaining
that it cannot #include <curl/curl.h>. It's unclear why this doesn't occur
with Go 1.7.6, although I think it may possibly be a race condition
because we DO every so often see a build failure for that same #include.
Regardless, it's harmless to add the CGO_INCLUDE_DIRS directive now and
it allows us to build with Go 1.11 on MacOS Mojave.

Change-Id: Ibdc193da194dcab9f4d1755b139415e2c9d040c5