MB-30907 Log error message when JSON unmarshall fails
When JSON unmarshall fails with errors like:
"json: cannot unmarshal string into Go value of type int",
it is difficult to root cause which entry in the input message
has a datatype mismatch. To debug these problems, a error
message is added when ever JSON unmarshall fails. The input
message is logged as a string

Change-Id: I38df7564170695f10045542deb6b19b90e78b92e