MB-34417 Avoid projector dcp_feed race condition during clean-up
When genServer() closes, it sets feed.conn to nil. This can happen
while doReceive() can read from conn (i.e. at feed.conn.SetMcdConnectionDeadline
or feed.conn.ResetSetMcdConnectionDeadline). This can cause a panic in projector

Instead of using feed.conn in doReceive(), we use "conn" object which was passed
to doReceive(). This way, even if feed.conn is set to nil, doReceive() will not
panic as it still holds a valid reference to connection

Change-Id: Icfb26265f93c1ce87aec40ad84b937ad90f54432