MB-32655 Propagate mutation arrival time from projector to indexer
The moving average value of mutation processing latency is propagated
every minute to indexer. The latency is calculated per stream per host.
Indexer adds its own latency to the projetor latency and returns the
overall latency

To identify from which host a mutation is being received, a map is
maintained (per bucket) between a vbucket and the host on which the
vbucket is active. This map is updated during stream_begin messages
and stream_end messages.

Change-Id: Ibe2da0bd9de4eb938e3445986a4236a00d74b1de