MB-34633 Disable new snapshot for vbuuid change
This change disables updation of HWT and generation of new snapshot
if there is a vbuuid change without any mutation.
It helps to reduce the possibility of rollback as memcached
can loose vbuuid in crash/restart loop while indexer can update its
HWT and generate snapshot for it. On restart of DCP stream with
those missing vbuuids, memcached could ask to rollback to 0.

Change-Id: I8ec94b2c446910e1208c3120574de5916d00ac21