MB-34821: Avoid clone for broadcastBootstrapStats
IndexerStats.Clone is a shallow copy where the clone will contain
pointers to the original stats. So, values set in the clone or original
will be visible in both. Therefore, this meant that values set in
broadcastBootstrapStats were visible in other clones. Since the values
of the other stats are filtered out in LifecycleMgr.handleNotifyStats,
it is ok for them to have default values.

Change-Id: I4db855277b0dd7c71fda68d1ccac8d3699cf8867