MB-35199 Fix concurrent map iteration and map write for fCmdGetStats message
In feed.go, when a request is made to get the feed level stats, the
feed.stats object is being updated and returned. This can cause
a concurrent map iteration and map write panic as it can happen
that one go-routine is clonning the feed.stats object while another
go-routine is updating the feed.stats object

This patch aims to fix the concurrent map access problem by initializing
a new FeedStats object whenever stats are to be retrieved from feed.go.

Change-Id: I45eac9efaec1065713a45f94ee35ecf1cfe990d1