MB-34579 Do not include DCP_CLOSESTREAM message for buffer-ack accounting
Including DCP_CLOSESTREAM message size in accounting for sending buffer-ack
message can lead to over-counting of data from projector. When a buffer-ack
is sent with this size included, it can lead to ThrowExceptionUnderflowPolicy
from DCP producer. This exception leads to closing the connection thereby
effecting all the streams shared by the connection.

As per DCP documentation, DCP_CLOSESTREAM should not be included in accounting
for sending the buffer-ack message

Change-Id: I9508d10de29b6592e3db90e344bbbab0b02d1445