MB-35701 Set default value of isBalanced to true
In rebalance_service_manager, the default value of isBalanced flag is set
to false. If the indexer process restarts after rebalance is done, this
flag will be set to false and the rebalance button is activated assuming
that index service requires rebalance. This is not a expected behaviour.

The default value of isBalanced is set to true so that indexer informs
ns_server about rebalance only in some special scenarios

Change-Id: Iaf0d520ce7e4270d2e1be17cf065d97b742b0345