MB-32633: Add new index stats for array index
New stats added are:
- docid_count: Number of docids that are currently indexed
- avg_array_length: Average number of items in each array

New field in IndexStats to track docid_count. This is set
in Slice.Statistics which is called for gathering storage stats.
The main and back store items count is used for primary and
secondary indexes respectively.

For MOI's NodeTable, expose ItemsCount as the sum of fastHTCount
and slowHTCount. Added a unit test for the same.  Calculate back
store's total items count by summing across workers.

Getting info from forestdb can fail, so avoid updating the stats
in such cases.

Change-Id: I9bdb818290c43b82bfbe8f4463674c3a55464528