MB-35699 Cleanup book-keeping in upr.go on STREAMEND
Feed book-keeping methods (addToFeed, removeFromFeed) in upr.go expects
that a vb is mapped to only one memcached.dcpFeed. During rebalance, when
memcached issues a STREAM_END but the vb-book keeping is not cleared in
upr.go. This can result in one vb mapped to multiple memcached.DcpFeed's.

This patch addresses the issue by cleaning up book-keeping in upr.go when a
StreamEnd is received from memcached

Change-Id: I8d717bdb956047ac2f70878375e0ddda7971985d