MB-35815: Reset memdb stores in case of recovery from disk corruption
In case of MOI, disk corruption is identified only during loadSnapshot.
If the latest snapshot is identified as corrupt in disk layer,
reset the memdb stores before attempting the loading of previous
snapshot.

Also, in case of error in loadSnapshot, close the corresponding
snapshot to clear the in memory stats and references.

Change-Id: I501cf2d9c474b3741f063a196768617a54eb61ec