MB-35328 Handling KV failover from indexer
When a node is failed over, it can take arbitrarily long time for
ns_server on the failed over node to realise the failover and act
on it. When it is a KV node that is failed over, in some cases
(E.g. indexer is ahead of replica and ns_server takes more than
20 minutes to realise the failover), this can lead to a rollback to
"0" for indexer

The rollback situation can be avoided to some extent by making indexer
to act pro-actively on KV node failover rather than waiting for ns_server
of the failed over node to terminate the projector process/ intimate
memcached about failover

Change-Id: I4602f95f205d3a483c9bbdd550f5ffcb67c37a6c