MB-35879: Handle error from CreateRecoveryPoint
Rollback and CreateRecoveryPoint can be called concurrently. This
can result in creating recovery point on snapshot that has been
created before rollback. If only one of mainstore and backstore recovery
point creation succeeds, then clean up the other. Additionally, remove
a recovery point if it is present in only one of the stores.

Change-Id: I47f64827d85c1635c2f24b7574e95e11735fa12e